the example code was wrong?

+++ bot.py
@@ -5,28 +5,33 @@ import discord
 with open("config.toml", "rb") as f:
     config = tomllib.load(f)
 
-
-class MyClient(discord.Client):
-    async def on_ready(self):
-        print(f"Logged on as {self.user}!")
-
-    async def on_message(self, message):
-        print(f"{message.author.id}")
-        try:
-            if message.author.id == 1259617744379183124:
-                await message.delete()
-                msg = discord.Embed(
-                    color=(235, 64, 52),
-                    title="SUNDANCE MESSAGE DETECTED: BUSTER CALL INITIATED",
-                )
-                msg.set_image("https://tenor.com/view/buster-call-ohara-gif-27694097")
-                await message.channel.send(embed=msg)
-        except Exception as e:
-            print(f"exception: {e}")
-
-
 intents = discord.Intents.default()
 intents.message_content = True
 
-client = MyClient(intents=intents)
+client = discord.Client(intents=intents)
+
+
+@client.event
+async def on_ready():
+    print(f"Logged on as {client.user}")
+
+
+@client.event
+async def on_message(message):
+    if message.author == client.user:
+        return
+
+    try:
+        if message.author.id == 1259617744379183124:
+            await message.delete()
+            msg = discord.Embed(
+                color=(235, 64, 52),
+                title="SUNDANCE MESSAGE DETECTED: BUSTER CALL INITIATED",
+            )
+            msg.set_image("https://tenor.com/view/buster-call-ohara-gif-27694097")
+            await message.channel.send(embed=msg)
+    except Exception as e:
+        print(f"exception: {e}")
+
+
 client.run(config["token"])
